Fix bug in manual indicator mode

+++ src/zigbee/relay_cluster.c
@@ -27,7 +27,7 @@ void sync_indicator_led(zigbee_relay_cluster *cluster);
 zigbee_relay_cluster *relay_cluster_by_endpoint[10];
 
 void relay_cluster_callback_attr_write_trampoline(uint8_t endpoint,
-                                                  uint8_t attribute_id) {
+                                                  uint16_t attribute_id) {
   relay_cluster_on_write_attr(relay_cluster_by_endpoint[endpoint],
                               attribute_id);
 }
